refactor replay into two files
